Move scoring functions to utils/metrics.py
